doc_scraping/requirements.txt
openai
pandas
numpy
faiss-cpu
requests
beautifulsoup4
python-dotenv